refactor(testing): split app/tests.rs into domain-separated files
45 tests extracted from the 1,626-line monolith into 5 targeted files:
  navigation_tests.rs  (9)  — pane_navigate, send_to_pane, boundary nav
  focus_tests.rs       (13) — focus history, back/forward, stale tile guard
  notification_tests.rs(17) — snooze, auto-dismiss, quicknote preemption
  context_tests.rs     (10) — child context, portal depth, cascade delete
  misc_tests.rs        (1)  — spawn_pane window targeting

No test logic changed. All 45 pass. Run targeted with:
  cargo test "app::tests::navigation"
  cargo test "app::tests::focus"
  cargo test "app::tests::notification"
  cargo test "app::tests::context"
