+++ src/app/tests/mod.rs
@@ -1,0 +1,5 @@
+mod navigation_tests;
+mod focus_tests;
+mod notification_tests;
+mod context_tests;
+mod misc_tests;
